관리자페이지 구현 완료(통계남음)

+++ src/main/java/com/meongnyang/shop/dto/request/admin/ReqModifyProductDto.java
@@ -26,7 +26,7 @@ public class ReqModifyProductDto {
     private Long categoryId;
     @NotNull
     private Long productPrice;
-    private Long productPriceDiscount;
+    private int productPriceDiscount;
     private String productDetail;
     private String productBrand;
     private String productModel;
@@ -46,7 +46,7 @@ public class ReqModifyProductDto {
     private Long arrivalQuantity;
     private Long minAlertQuantity;
     private Long alertSetting;
-    private Long outOfStock;
+    private int outOfStock;
 
     public Product toEntity() {
         return Product.builder()
